+++ app/src/main/res/layout/layout_upload_item.xml
@@ -4,17 +4,37 @@
     android:layout_width="match_parent"
     android:layout_height="50dp">
 
-    <androidx.cardview.widget.CardView
+    <FrameLayout
+        android:id="@+id/back"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:cardCornerRadius="10dp"
-        app:cardElevation="0dp">
+        android:background="@color/red">
 
         <ImageView
-            android:id="@+id/img_preview"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_gravity="end"
+            android:layout_marginEnd="20dp"
+            android:src="@drawable/ic_delete" />
+    </FrameLayout>
+
+    <FrameLayout
+        android:id="@+id/fore"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/white">
+
+        <androidx.cardview.widget.CardView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:scaleType="centerCrop" />
-    </androidx.cardview.widget.CardView>
+            app:cardCornerRadius="10dp"
+            app:cardElevation="0dp">
+
+            <ImageView
+                android:id="@+id/img_preview"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:scaleType="centerCrop" />
+        </androidx.cardview.widget.CardView>
+    </FrameLayout>
 </FrameLayout>
